ai_assistant2.0/requirements.txt
# Core
pyttsx3
speechrecognition
requests
webbrowser

# Gemini AI
google-generativeai

# GUI
pillow

# App/Window Control
pygetwindow
pywin32

# Automation
pyautogui
pyperclip

# Optional (for News, etc.)
newsapi-python

# For Windows only (used in some modules)
pywin32